Remove crossorigin attribute for CSS
In a bid to improve behavior on Safari. Really this shouldn't make a
difference, since crossorigin="anonymous" is a no-op when it's same
origin, but maybe Safari is not doing things correctly.

Safari users often report seeing a completely unstyled page, seemingly
because the Bootstrap CSS doesn't load. The problem is that credentials
must be sent to load the CSS, since the entire /213/examreg/ path is
protected by Shibboleth.